+++ tests/testthat/test-output-batches.R
@@ -1,0 +1,11 @@
+patrick::with_parameters_test_that(
+  "sink_batches works",
+  {
+    lf <- as_polars_lf(mtcars)
+
+    expect_snapshot(
+      lf$sink_batches(\(df) print(df), chunk_size = 10, maintain_order = TRUE, engine = engine)
+    )
+  },
+  engine = c("streaming", "in-memory")
+)